requirements.txt
streamlit
google-generativeai
edge-tts
moviepy
manim
pymupdf
python-dotenv